Check for arch-specific wheels when copying generic wheels
Some wheels are incorrectly detected as generic by
`wheel_architecture.sh` because they have no dependencies on external
packages but contain C extensions or bundled libraries that use SIMD.

To detect this, `cp_wheels.sh` now checks for existing wheels (same
package name) in the wheelhouse that are in arch-specific locations.
If any are found, `cp_wheels.sh` refuses to copy the wheel to a
generic location. The architecture can be specified with the new
`--arch` option.