Pin torch requirement based on libtorch. Tag wheel if needed.
A wheel can depend on torch and/or build on libtorch. Therefore, we pin torch version when libtorch is linked.
Otherwise, the wheel needs torch at runtime, for instance:
- torch
- torch>1.10
- torch<1.14
and we don't want to pin a version in these case since it does not build on libtorch.

We want to tag the wheel only when it would be overridden when building with a new torch version.