Remove `scipy` from PYTHON_DEPS_DEFAULT
With Update/default numpy #77, we use `oldest-supported-numpy` module instead of installing `numpy` in the virtual environment.

By default, we were installing the latest `scipy`. Since version `1.10` needs `numpy<1.27.0,>=1.19.5`, pip installs a `numpy` that is greater than the `oldest-supported-numpy` for python 3.8 and 3.9 (1.17.3, 1.19.3 respectively), resulting in the wheel's numpy requirement being adjusted to 1.24 since it built with this version.